requirements.txt
pyserial>=3.5
flask>=3.0
flask-socketio>=5.3
eventlet>=0.36
